Move prep_tokenizer to tokenizer_utils.py
